requirements.txt
django==3.0
djangorestframework==3.10.0